+++ src/main/java/ru/sharanov/teacherservice/repositories/TeacherRepository.java
@@ -8,7 +8,7 @@ import ru.sharanov.teacherservice.model.Teacher;
 import java.util.List;
 
 @Repository
-public interface TeacherRepository extends JpaRepository<Teacher, Long> {
+public interface TeacherRepository extends JpaRepository<Teacher, Integer> {
     @Query("SELECT t FROM Teacher t JOIN t.student s WHERE s.id = :studentId")
     List<Teacher> findByStudentsId(Long studentId);
 }
